Actually move the player, Improve TurnMenu.
For the TurnMenu, the player and location (via the player class) are provided so more information can be shown in the menu. For now, the location name is added.

The player is also moved when the move command is issued. This sets the player.location to a Location object, so it can be more easily referenced for that player's actions (instead of calling Game.get_location() often to get the actual Location object).

This closes #13.