Add JobMenu and job_apply command.
The JobMenu will handle any job functions like applying for a job or working a unit of time. Right now it only really handles the applying for a job.

I also added the job_apply command in the Game class. This applies the user for the job, which will always accept at this point and give the player the job.

There was also a small addition to Location to get a job by symbol.