Attempt to sort job application list.
When the player applies for a job at a location, the job list is displayed through the JobMenu class. I tried to sort this list by Job.rank, but it doesn't seem to be working. When it is working, the list for the bank location should be:
t. Teller
m. Manager
(Enter to cancel)