Add location variable to Job class.
To reference the location that holds a job, there is a location variable in the Job class which holds a Location object (or may hold None, but this should not be possible).

I also modified the player info display to include where the location is of the job that the player currently holds. This closes #56.