Display a list of purchasable items to the user.
This adds a ListMenu, which is a generic menu that can be used to display a list of objects and returns the actual object that the user selects. This is different than other menus, which return the option the user selected (like 'm' for move).

The 'buy items' menu option should only show when there are items at a location.