Added BuyMenu to show the prices for each item you want to buy.
I couldn't find a way to use the ListMenu but to show the prices next to the item, since the string representation of Item is used elsewhere, so I didn't want the price to show up when you went to use the item. This seems to be the best option for now, although it could certainly stand to be redone if we find a better way.