Fixed some areas to correctly use the new attributes dictionary.
Particularly in the game class, I had to change how money was being taken. It is not very intuitive for the programmer, but I think this will be more expandable down the road (which is yet to be seen).

I also changed the few items currently in the map to use just the major attributes, and tested with the book in the library.