Adjusted costs in map file to use negative numbers.
This should be more intuitive when making the map file, so any negative value will take that away from the attributes (time, money, happiness) and positive values add to it.

I also used abs() when checking if the player has enough money or time. I did NOT change how time works (or is defined in the map file) yet, but that may move to a major attribute to work like everything else.