Fixed all game commands to use the Time attribute in the Player class.
This will be better in keeping this with the player rather than a variable in Game, especially for multiplayer games down the road.

I changed the order of checking for valid parameters of a command before checking to see if there was enough time, which makes more sense, and is necessary.

I also changed the map file to show negative values for time. This also means that using an item can use time now by simply putting it in the effects list for that item (rather than a code change).