Added a step in Game.new_turn() to move all players to the start_location.
This adds a function in Map to get the start location, as well as the loading of the start_location attribute from the map json.

There is also an addition to the Location object to mark it as the start location, as defined by the instance variable start_location.

Fixes #85